Bug fix: No longer freezes when parsing /\S/, /\D/, /\W/, /\H/
This was being caused by the way that the parser is designed to work
with character groups, e.g. /a-z/:

Since \S was being defined as [ ... , ",", "-", ".", ... ], the parser
kept trying to expand the "range" from "," to "." - which included a
"-", so it got stuck in an infinite loop!

The easiest solution was to ensure that the definition of /\S/ etc.
always put "-" at the START of the array, so that the parser would
handle it correctly (since it will then treat it correctly, as in
/[-abc]/