Added some very rigorous specs for named properties
These tests are SLOW(!!!)
This show it's important I need to change the way things are being
stored/calculated in order to boost performance...

I will now also be able to use these specs as a tool to benchmark
perforance, which is nice.
Currently takes about 1 minute on my machine; it SHOULD be no more than
maybe 2-3 seconds.