Refactored OrGroup to fix probability distribution bug
Previously, if you had a regex like: `/a|b|c|d|e/`, then the code would
have parsed it as strictly BINARY "or" groups, i.e.

`/(a|(b|(c|(d|e))))`#

For the application of Regexp#examples, this was fine. But for
Regexp#random_example, it introduced the issue that it's now MUCH more
likely for "a" to be randomly chosen than "e".
In fact, "a" has a 50% change whereas "e" only has a 6.25% chance.

This patch fixes the issue, by changing the way that the regex is parsed
- i.e. I have removed the "BINARY" attribute of the OrGroup, so it will
now just be parsed as:

`/(a|b|c|d|e)/`